Updates `file_path_of()` to return paths relative to the main workspace directory inside the execroot.
This means a target in the main workspace such as `//path/to/internal/pkg:target` would return `./path/to/internal/pkg/target` while a target from an external workspace such as `@wksp//path/to/external/pkg:target` would return `../wksp/path/to/external/pkg/target`.

This is because actions typically execute in the main workspace of the execroot, so the directory structure would look like:
* execroot/
    * __main__/ # <-- Working directory.
        * path/to/internal/pkg/target
    * wksp/
        * path/to/external/pkg/target

This path structure means that the output of `file_path_of()` can be used to directly reference any internal or external file in an action based on the default working directory.