Bumps `@aspect_rules_jasmine` dependency to current latest.
Refs #48.

I actually want to bump the `@aspect_rules_ts` dependency, but this appears to require `@aspect_rules_js` to be bumped as well, and there seems to be an unintentional breaking change when used with `@aspect_rules_jasmine`, so we need to upgrade this first.

See notes in https://github.com/aspect-build/rules_js/releases/tag/v1.17.0.