Removes `css_map()` and inlines its implementation in `css_bundle()`.
Refs #46.

The import map is only ever generated as part of `css_bundle()`. It needed to be a different rule before because `postcss_multi_binary()` was a macro and needed to be called at that level. But now that `css_bundle()` is a custom rule, we can generate the import map directly, which is much more straightforward.