Renames `@npm` to `@rules_prerender_npm`.
We can't generally claim the workspace name `@npm` since it is very likely that user code will want to declare that workspace for itself. Instead, we use `@rules_prerender_npm` which installs the `package.json` in the `@rules_prerender` repository. Users can define their own `@npm` which installs their own `package.json`.

Unfortunately this means that local development in this repository is a bit confusing since the `@npm` workspace doesn't exist, though this is a small price to pay.