Moves `//packages/renderer/...` to `//tools/binaries/renderer/...`.
Refs #67.

This is because the renderer is invoked as a tool, not installed as a package by the user. This is a slight misnomer as `//tools/binaries/renderer/...` does not export a `js_binary()`, however it does export a library which gets wrapped into a `js_binary()` under the hood, and treated as _effectively_ a tool, even if technically it isn't.