Updates script bundling to use multiple entry points.
Refs #26.

Previously, we collected all the scripts referenced by `includeScript()` and bundled them together into a single JavaScript file which was injected into _every_ HTML page. This is unnecessarily broad, as not all pages use all the JavaScript. Instead, we should bundle the JavaScript for each HTML page separately, with shared chunks for common functionality.

This makes a few updates to support multiple entry points.
1. It updates annotation extractor to no longer generate a list of included scripts, but instead generate a map of HTML file -> list of included scripts. This provides enough information to know which scripts are associated with which page.
2. It updates script entry generator to generate a directory (`TreeArtifact`) of entry points instead of a single entry point. This directory contains an entry point for every HTML file which includes at least one script.
3. It replaces `rollup_bundle()` with a custom rule which invokes the `rollup` binary under the hood. This is necessary because `rollup_bundle()` supports multiple entry points _only_ if those entry points are known at analysis time. In this case, they are not because we can't know what HTML files will include scripts until execution time.
4. It updates resource injector to no longer accept a single JavaScript bundle, but instead accept a directory of JavaScript bundles. These bundles are directly copied to the output directory and also any matching HTML file gets its associated JavaScript file injected.

In the process I also updated `rollup.config.ts` so it is authored in TypeScript and includes its own plugin dependenices.