Adds `FileSystem` and fake implementation.
Refs #33.

This is useful to mock and stub the file system without relying on spying on imports in Jasmine, which doesn't work with ESM.

I investigated a few existing "fake file system" dependencies, however I wasn't really happy with how any of them worked or the supply chain implications of using them. I'm hoping to keep file system usage simple enough that it won't grow to be a maintenance headache, but we'll see how things shake out.