Deletes `fs` module.
Refs #33.

This module's entire purpose was to proxy the underlying Node.js `fs` module in a way which could be spied upon. However, in native ESM we can't spy on even user modules which are imported. We already have `FileSystem` and `FileSystemFake` as a replacement and all tests were successfully migrated to it, so `fs` is no longer needed.