Updates and fixes ESLint configuration.
ESLint was complaining that it was too out of date for the currently installed TypeScript version. Bumped the `eslint` and `@typescript-eslint/*` versions to be current latest. This did affect two existing lint checks, one I didn't really care about in the first place (`@typescript-eslint/ban-types` was unnecessarily banning `object` IMHO), though the other seems like an actual regression in `@typescript-eslint`. I don't care that much about this check in particular, so just dropping the ignore for now.