Updates `includeScript()` to use a relative path with `import.meta`.
Refs #69.

This greatly simplifies imports since most scripts are adjacent to the prerender logic where they are included. The path is resolved immediately and validated. Relative paths _must_ include a `*.js`, `*.cjs`, `*.mjs` file extension because checking this in Rollup is tricky and far away from the actual developer mistake. Bare imports are currently disallowed as I haven't tested or evaluted there behavior. Longer term they should probably be supported and resolved to `node_modules/`, but for now they are disallowed.