Updates `inlineStyle()` to use relative imports with `import.meta`.
Refs #69.

This brings `inlineStyle()` into parity with `includeScript()`. Input strings are relative to the file they are imported from.

I had to update the CSS import map to drop the workspace name in order to use `wkspRelative()` as is. I think this is ok for now as we don't really have a story for inlining styles from different workspace. Most likely this would be done by using bare imports and inlining files from NPM packages, rather than external workspaces.