Updates all `includeScript()` calls to include file extension.
Refs #69.

This is more consistent with ESM JavaScript imports. Unfortunately Rollup seems to automatically add file extensions by default, and I can't find a way to turn it off. It might seem like `@rollup/plugin-node-resolve` is responsible for this behavior, but it actually isn't. That plugin only resolves plain imports to `node_modules/`, Rollup itself resolves `./foo` to `./foo.mjs`.

I tried to add a custom plugin which errors on missing file extension, but the `resolveId()` callback never seemed to be invoked for reasons I don't understand.