Removes `async` from injected scripts.
I didn't fully understand the meaning of `async` when I first added it and thought it was helpful for improved performance. However after [learning more about it](https://blog.dwac.dev/posts/html-fragments-routing/#router-loading-timing) I don't think it's necessary here. The original goal was just to set up the user for good performance by default and not block rendering, however using `type="module" async` actually makes the script execute sooner if it happens to load faster than the page is downloaded. This can be useful in some situations, but is quite rare in practice, especially since `@rules_prerender` is all about prerendering all the HTML content which could be hosted on a CDN and will very likely download the HTML first unless you have a massively large HTML page.

Using `async` does mean that the JavaScript must expect that the page is not fully parsed and wait on `DOMContentLoaded` if necessary, which greatly complicates the implementation. Removing `async` means scripts are always executed after the HTML is fully parsed (because `type="module"` is implicitly `defer`). This is much easier to work with and doesn't really hurt performance.

Occasionally `async` is useful and we might add an option for it in the future. However the simplicity of _not_ being `async` is worth using that as the default behavior.