Refactors `WORKSPACE.bazel` to reuse setup for external repositories.
This reduces duplication between `WORKSPACE.bazel` and user `WORKSPACE.bazel` files, reducing divergence.

I also discovered that the `@bazel_skylib` declared wasn't actually used because `@aspect_bazel_lib` declared a newer version first and we were dependent on that version. Moved the declaration first so it is used instead and updated it to the version used by `@aspect_bazel_lib`.