Adds `SafeHtml` module.
This represents a "safe by construction" object, similar to `TrustedHTML`. When given an instance of type `SafeHtml`, it's content can be trusted and rendered to the end user. Currently there are two ways to construct a `SafeHtml` object:
1. `safe` - A tagged template function which asserts that no arguments are given and therefore the template is a hard-coded constant.
2. `unsafeTreatStringAsSafeHtml()` - A function which takes unsafe string content and constructs a `SafeHtml` object from it.

The latter option is named to be clearly unsafe and performs no sanitization on its inputs. This is intended to be used only on the output of other trusted template engines such as `lit-html`.

I deliberately kept the `SafeHtml` class private so it is not easy to get a reference to it (though still possible) and froze all the relevant symbols to try to limit potential vulnerabilities. I also exported the unsafe function from a separate file with a separate Bazel visibility list, which should hopefully help mitigate its usage. Unfortunately it still needs to be exported from `rules_prerender`, but hopefully the name is enough to make people uncomfortable with using it.