Adds `PrerenderResource.fromText()`.
Refs #71.

This move raw string support in `PrerenderResource` to a different function. The main objective is to make it significantly harder to accidentally write a raw string to an HTML file, given that there are little to no safety guarantees. `PrerenderResource.fromText()` accepts a raw string, but refuses to associate it with a `*.html` or `*.htm` file. This supports writing textual files which are _not_ used as HTML content. Instead, users are directed at `PrerenderResource.of()` and `SafeHtml` which should be preferred.

Of course no protection is perfect here. Users could write to a `.definitelyNotAnHtmlFileISwear` and then rename it later in the build process. Though the script and style injection tooling is limted to `*.html` files, so they'd be throwing away a lot of functionality to do that.