Adds `@rules_prerender/lit_engine`.
This package supports Lit as a templating engine with strong XSS protections. I needed to use `@lit-labs/ssr` in order to actually render the template to a string, and it's a bit unfortunate that it is still experimental, but I'm not sure there's a better approach. I suspect the "render to string" part should be pretty stable and most of the complexity is around support server-side custom elements (something we want `rules_prerender` to support anyways) and hydration (something we're planning to tackle completely differently). So I don't think this experimental dependency should cause too many issues in practice, but it is a risk to be aware of.

For now, rendering to `SafeHtml` is an async operation meaning it requires users to `await` (which they might not otherwise need to do) _and_ it doesn't take advantage of `@lit-labs/ssr`'s streaming capabilities. We'll probably want to follow up afterwards to address both of those.