Restricts overbroad `//visibility:public` usages.
`//visibility:public` actually makes the target available to any target _in any workspace_, meaning they are publicly visible to user workspaces. For most of these they were actually intended as being public _internally to `@rules_prerender`_. This updates as many targets as possible to be as restrictive as possible.

Unfortunately a decent number of these are referenced from macros and rules invoked in the user's workspace, so they need to be `//visibility:public` even when they are technically implementation details.