Requires `TemplateStringsArray` in `safe` tagged template literal.
This prevents passing in a regular array, which makes it more resiliant to malicious input. Even if an attacker managed to trick an application into passing an array of input into `safe`, it would be rejected for not containing the `.raw` property. Ultimately there are other ways of faking a `.raw` property, so this doesn't fundamentally solve the problem so much as making it significantly harder to compromise. Honestly the biggest benefit is that it makes it useless for real user code to invoke `safe` as a regular function, keeping more users on the intended tagged template literal approach.

Relevant thread: https://techhub.social/@develwithoutacause/109990391783832781.