Makes `//:node_modules/@rules_prerender/declarative_shadow_dom` private.
Refs #71.

I accidentally used this target today when I actually wanted `//:prerender_components/@rules_prerender/declarative_shadow_dom`. There is not much use for the Node module and using it directly violates the expectation that all the various "slices" of the component stay together in a cohesive way. Limiting visibility makes it easier to catch this mistake in the future.