Bumps `@aspect_rules_ts` to current latest version (`1.3.2`).
Also dropped the explicit dependency in `examples/external/`. There is technically a direct dep on `@aspect_rules_ts`, and in a real project it should probably pin a specific version, but as a test project in `@rules_prerender`, it's too easy to forget to update, so I just dropped it. That example will use the same `@aspect_rules_ts` version from the `dependencies.bzl` file.