Updates release process to generate release notes and include the SHA of the workspace.
This essentially tars the whole workspace and uploads it as a release artifact. This gives an opportunity to hash the workspace tar and include it in the release notes.

No way I got this right on the first try, but there's no good way to debug without just doing it.