Adds visibility check to `prerender_component`.
Refs #40.

This uses an aspect so `prerender_component` can inspect its component slices dependencies. The aspect provides the `visibility` attribute which `prerender_component` asserts on. The idea is that slices for a component _must_ be defined in the same package as the component _and_ must have private visibility. With both of these and assuming a structure where each `prerender_component` gets its own Bazel package, it should be impossible to accidentally depend on a component slice without going through the `prerender_component` reexports.

This implements option 4 in [this dicussion](https://github.com/dgp1130/rules_prerender/issues/40#issuecomment-1518516150).