Adds script-only dependency to rewrite example.
Refs #40.

This dependency edge is only through a client-side script, not the prerender code. A client side script is pulled in from a dependency component which reads a text file. This text file is only included in the output if `@rules_prerender` is able to find it correctly. This verifies the new functionality of `prerender_component2` and confirms that script and style-only deps are supported.