Removes `bundle_js` flag from `prerender_pages`.
This flag served only as a minor optimization for skipping the bundling phase when no client-side JavaScript was expected. It was required before because Rollup would throw a hard error if no entry points were given to be bundled. Now the `js_bundle` rule passes when given no inputs and generates an empty output directory as would be expected. As a result, `bundle_js` is no longer required to avoid errors and the optimization benefits are so trivial as to not be worth the trade off. If a user forgets scripts are disabled for a page and then adds scripts to a component, they may be very confused to find those scripts are not loaded. Requiring a bundle step for client-side JavaScript means this mistake can never be made.