WIP: Fork webtesting infrastructure from Angular dev-infra.
https://github.com/angular/dev-infra/tree/1ad20ef9dd457de967252283c1a968b0d702d0ae/bazel/browsers/

This seems to support a setup for manually versioning browsers in the repository. It still has a dependency on `@io_bazel_rules_webtesting`, but hopefully we can break that eventually.

This commit is a straight copy with *zero* modifications. The only difference is adding a `README` to describe where the fork came from. Any required changes will come in future commits.