Migrates `pnpm-lock.yaml` to version 6 and upgrades to latest `@aspect_rules_js`.
Installing pnpm@8 requires migrating the lockfile to version 6. Using this new lockfile and pnpm seems to break `@aspect_rules_js` as `preact-render-to-string` is no loaded correctly in `//packages/preact/...`. I suspect this is because the dependency is duplicated between the package dependency and the workspace dev dependency, causing `@aspect_rules_js` to do the wrong thing. Updating to latest seems to fix this, though I'm not able to pin down a clear reason why. The changelog for `@aspect_rules_js@1.24.0` indicates that it supports pnpm@8 and lockfile v6, but it appears to be broken for my case. Presumably some future bug fix has addressed this problem.

Also re-adding the Jasmine dependency which somehow got dropped in a previous commit.