Adds debug Chromium setup.
Refs #74.

This updates the Chromium browser to choose a forked configuration without `--headless` when in debug mode. This way the browser can be inspected during tests.

For some reason, `--//tools/browsers:use_debug_browers` needed to be enabled manually. I thought this enabled by `--config debug`, but apparently not. I've added it to `debug.bazelrc` to make this more automatic.