Upgrades Yargs to v17.
Refs #74.

This is to align with the version used in Angular Dev Infra's browser update tool. Also it's just a good upgrade to do.

The API changed slightly in Yargs, and we need to explicitly stay synchronous with `.parseSync`. I took the opportunity to standardize some of the formatting on the various tools. Also adds `.strict()` which I just discovered and which ensures no unknown commands are passed.