Removes unneeded code from the browser update script.
Refs #74.

A decent portion of this script was dedicated to pushing the browser binaries to a mirror owned by Angular. We don't have permission to do that here so it can be dropped.

I'm a little sad to drop the Firefox part, as it seems only Chrome is supported for pulling actual version numbers. I may try to revive that in the future if and when upgrading Firefox becomes an issue.