Links vendored `@io_bazel_rules_webtesting` workspace into the main workspace.
This requires copying most of the forked `WORKSPACE` file, since Bazel workspaces do not compose. I also _had_ to rename to `WORKSPACE` because apparently some of the Go tooling requires that name and throws an error for `WORKSPACE.bazel`.

This messed up the `pnpm test` script, which would no longer find the root workspace. Therefore we needed to update the script to allow `WORKSPACE` and `WORKSPACE.bazel` files, but ignore `third_party/` so we don't run all of `@io_bazel_rules_webtesting`.

The incompatible flags revealed a couple errors in the Go tooling required to support `@io_bazel_rules_webtesting` which needed those flags to be turned off temporarily.