Removes MacOS and Windows versions of browsers.
`bazel query //...` was broken because it would attempt to download and resolve the MacOS and Windows browser binaries. Firefox MacOS seems to have an error and not download successfully. Chrome Windows and MacOS seems to be fine, but neither are used so no need to waste effort downloading them. I left the MacOS and Windows browser repositories declared, just not depended on anywhere. That should make it easier to use them in the future when needed and also allow them to be updated alongside the Linux versions when the browsers are updated.