Adds `netlify_deploy` macro and sets up docs site deployment.
The `netlify_deploy` macro is a fairly straightforward wrapper around `@netlify/cli`. The biggest challenge was installing it, as it does not currently work with pnpm. I was able to work around this via `pnpm.packageExtensions` and I have a proper fix for the dependency [in progress](https://github.com/netlify/build/pull/5247/).