Adds 404 not found page.
Also adds a Netlify configuration file which configures the not found page to be served for any unknown path.

I took the time here to set up a new `www/` directory to serve as a kind of "file-based routing" setup. It's not _really_ file based routing, but I think it's a good convention to use for project structure.