Adds `<InlinedSvg>` to `@rules_prerender/preact`.
This required some refactoring to perform relative path resolution using `import.meta` like `includeScript` and `inlineStyle` do. This improves performance and changes the styling story of SVGs, as they respect or don't respect certain styles based on whether they are inlined, loaded via an `<img>`, or embedded through an `<object>`, `<embed>`, or `<iframe>` (see https://css-tricks.com/scale-svg/).

I opted to use `readFileSync` for now to avoid making everything `async`. It's annoying and makes me feel dirty, but it's not the end of the world. In the future, it might be worth looking into adding some kind of "inlining" pass to the bundler, so `<InlinedSvg>` could emit an annotation which gets processed by the bundler to inline the file at the right location. Not sure if that's worth the effort right now.