Adds `CustomElement` type.
This provides the `defer-hydration` attribute to make it easy to use in a Preact context.

I'm a bit torn on this type because not every custom element will implement `defer-hydration`. Arguably it should be up to each individudal component to declare whether or not they support `defer-hydration`. However, if libraries like HydroActive and Lit do it out of the box, then _most_ components should align with this and realistically most of them won't actually declare than they do support it. This feels like the best compromise for now, but I may want to re-evalute in the future.