Adds `<Template />` to `@rules_prerender/declarative_shadow_dom/preact.mjs`.
This is a replacement for the `<Template />` element in `@rules_prerender/preact` but with the notable difference that it automatically includes the declarative shadow DOM polyfill when `shadowrootmode` is set. This removes the need to remember to do this for every component.

This only works because scripts are loaded in the same order they are injected into the HTML (when they don't import each other). As a result, as long as the DSD polyfill is included first, it will be executed first. `<Template />` enforces that a little better than users can, but ultimately we are relying on that ordering to do the right thing. I'm not sure happy with that, but I think it is the nature of polyfills to a certain extent, especially when the dependency on DSD is the prerendered HTML, not the client-side JS where an `import` would solve this problem. For now, I think this is the best we can do.

`<Template />` from `@rules_prerender/preact` still exists for now, but will be removed after all internal usages of it are migrated off.