Adds markdown rendering of the getting started page.
Refs #16.

This got a little involved, but I learned a lot in the process here. The general pipeline is:
1.  Perform an `async` read of the `getting_started.md` file.
2.  Parse the markdown with `gray-matter` to separate the frontmatter and markdown body.
3.  Parse the markdown body with `marked` and transform it into HTML.
4.  Validate the frontmatter with a `zod` schema.
5.  Render the page in Preact.

This relies on a "fetch-then-render" approach, where all the data is loaded up front asynchronously _and then_ rendered afterwards. I initially tried a "fetch-as-you-render" approach with `<Suspense />` but found it to be a bit lacking. Preact's `<Suspense />` implementation is still experimental, and currently intended for lazy loading components. Data loading with `<Suspense />` is not really a well-trodden path and does not have defined primitives, either in Preact or React. Most React frameworks seem to define their own data loading primitives which I was initially hoping to avoid, letting users structure their own rendering process.

For now, the easiest option is to fetch data up front and render afterwards. This isn't great performance and I'm not totally satisfied with it, but it's good enough for now. In the future, hopefully we can revisit async Preact and come away with a better path forward.

[See discussion in `preact-ssr-prepass`.](https://github.com/preactjs/preact-ssr-prepass/issues/55)