Adds initial getting started and `prerender_component` tutorials.
These are *very* rough drafts and include a lot of placeholders. The set up is way more complicated than it should be and I would love to simplify it via an `npm init @rules_prerender` feature. The component tutorial is only partially done as it only covers prerendering HTML. The markdown itself is completely unstyled and not pretty to look at. However this should give some initial content to start building the site around.