Fixes `<code>` overflow.
This was a tricky one which led to a few random changes. The problem was that when the viewport got too small, the "page" in `main` would shrink its width as expected. However, even a single long line of `<code>` would block this shrinking. `main` could not be thinner than the longest line of `<code>` inside it. This put a horizontal scrollbar on the `html` and look generally awful.

This sounds simple and I expected `pre { overflow-x: auto; }` to fix this issue, however I found that was not sufficient. This commit contains a couple fixes and one hack that I don't understand to address the problem:

1.  Add `pre { overflow-x: auto; }` - This doesn't do anything on its own and is not sufficient to solve the problem, but still necessary for the solution.
2.  Remove `rp-layout { height: 100%; }` - This is necessary to let it flow as tall as it needs to be, otherwise the `footer` is pulled up to always be in the viewport, which is not desired.
3.  Remove `body { height: 100vh; }` - Since we're accepting that the grid extends beyond the viewport, there's no reason to constrain `body` to the viewport either, though removing this isn't strictly necessary it makes sense to do so.
4.  Move `main { margin: x auto; }` to its child - Since `main` *is* a grid cell, it does not make logical sense to center the cell itself. Instead we need to center the cell's *child*.
5.  HACK: Add `main { overflow-y: auto; }` - There's a long comment about this and I don't understand why it necessary, given that `main` doesn't overflow and does not / should not present a scrollbar. However this one style is apparently critical to getting `pre { overflow-x: auto; }` to work as expected. Why this style has an effect like this on a descendent, I do not understand.

See https://stackoverflow.com/a/77413468/3995712 for more context.