Fix main content centering and box-shadow.
Since the `main` element no longer has `margin: 0 auto;`, child elements need to be centered properly for wide viewports. The gutter on the grid was covering the box shadow so it needed just enough margin to keep the box-shadow visible and needed to be a different element from the centered one.