Adds `testId` utility.
This is a small utility for generating `test-id` attributes _only_ when in a test environment, not for production builds. While this is starting as a docs-only utility, I'm considering including it in `@rules_prerender` proper. Some open questions to consider:
1.  Should this be restricted to `testId`, or should it apply to any kind of "test-specific" logic? Should it be named `testonly('foo')` and just return whatever value it's given or `undefined`? What about `if (isTesting) /* ... */`?
    *   Personally I think putting test only logic inside production code is kind of an anti-pattern to be avoided, with test IDs being one small situation where it is worth the trade off. I'd be interested to see if other clear uses cases arise.
2.  How to enable test IDs? Currently this requires a manual `enableTestIds` call, but if the API were promoted to the framework-level, it would be cool to enable it by default. Unfortunately this still runs a build time, not test time, since it needs to be used in the binary running `prerender_pages` for an `_test_cases.tsx` file. That means we can't automatically enable this logic during test-time execution. We could infer from `testonly` or put an option on `prerender_pages`, but I don't know if either is the right approach.

For now, I made this a module-level value because that presented the best API. It did require an internal "disable" functionality just to test in a stable manner. This is kind of awkward, but alternatives like putting the state in a class and instantiating that class are too unergonomic to use directly. This could be mitigated with a separate file which creates a singleton of this class at the top-level scope, but at that point the API I would be testing and the API users would be consuming would be far enough apart that I don't think it's beneficial to the design.