Adds an initial routing implementation to the docs site.
This presents an API for defining a route configuration which supports hierarchical routes and associates a `render` function with each. This makes it easy to define a large list of routes and render them individually.

The main feature here is that the route config is processed into a route forest which is passed into the `render` function as an input. This means every route knows what all the other routes are and can use this information to render site navigation.

While this API is limited to the docs site, it is entirely possible it will be promoted to a "real" public API in `@rules_prerender` in the future.